\title{IsoGeneGUI Graphical User Interface for the IsoGene package}
\name{IsoGeneGUI-package}
\alias{neaGUI-package}

\docType{package}

\description{
neaGUI is an easy to use R package developed to perform the network enrichment analysis (NEA) proposed by Alexeyenko et al. (2012). The NEA method extends the overlap statistics in GSEA to network links between genes in the experimental set and those in the functional categories by exploiting biological information in terms of gene interaction network. 
}


\details{
The Network Enrichment Analysis Graphical User Interface (neaGUI) is a user friendly interface for network enrichment analysis (NEA).
}

\author{
Author: Setia Pramana

Maintainer: Setia Pramana <setia.pramana@ki.se>
}
\references{

}

\keyword{package}
\examples{
\dontrun{
library(neaGUI)
neaGUI()
}
}
