\title{neaGUI }
\name{neaGUI}
\alias{neaGUI}
\alias{neaGUI-package}

\description{
This function will load the neaGUI package.
}


\usage{
neaGUI()
}


\details{
To run the package, we use the function: neaGUI().
}
\author{
Author: Setia Pramana

Maintainer: Setia Pramana <setia.pramana@ki.be>
}


\keyword{documentation}
\examples{
\dontrun{
library(neaGUI)
neaGUI()
}
}

