\title{neaGUI }
\name{neaGUI}
\alias{neaGUI}
\alias{neaGUI-package}

\description{
The neaGUI package was developed to perform the network enrichment analysis (NEA) proposed by Alexeyenko et al. (2012). The neaGUI requires the following R packages: nea, tcltk, KEGG.db, GO.db, reactome.db, org.Hs.eg.db, AnnotationDbi, and hwriter. 
To load the neaGUI package use the following command.
}


\usage{
neaGUI()
}


\details{
To run the package, we use the function: neaGUI().
}
\author{
Author: Setia Pramana

Maintainer: Setia Pramana <setia.pramana@ki.se>
}


\keyword{documentation}
\examples{
\dontrun{
library(neaGUI)
neaGUI()
}
}

